// src/constants/enums/role.enum.ts
export enum Role {
  ADMINISTRATOR = 'ADMINISTRATOR',
  ADMIN_COLLECTION = 'ADMIN_COLLECTION',
  ADMIN_BLOG = 'ADMIN_BLOG',
  CREATOR = 'CREATOR',
  ADMIN_MARKETPLACE = 'ADMIN_MARKETPLACE',
  ADMIN_NFT = 'ADMIN_NFT',
  ADMIN_LAUNCHPAD = 'ADMIN_LAUNCHPAD',
  ADMIN_USER = 'ADMIN_USER',
  VIEWER = 'VIEWER',
}